rocketmq-gateway/python_client/requirements.txt
grpcio==1.70.0
grpcio-tools==1.70.0
protobuf==5.28.3
